+++ jason/src/peer/media/mod.rs
@@ -309,7 +309,6 @@ impl MediaConnections {
                 }
                 Direction::Recv { sender, mid } => {
                     let recv = Receiver::new(
-                        inner.peer_id,
                         track.id,
                         &(track.media_type.into()),
                         sender,