+++ jason/tests/api/connection.rs
@@ -1,9 +1,10 @@
 #![cfg(target_arch = "wasm32")]
 
-use futures::channel::oneshot;
-use medea_client_api_proto::{
-    Direction, MediaType, PeerId, Track, TrackId, VideoSettings,
+use futures::{
+    channel::{mpsc, oneshot},
+    StreamExt,
 };
+use medea_client_api_proto::{PeerId, TrackId};
 use medea_jason::{
     api::{ConnectionHandle, Connections},
     peer::RemoteMediaStream,
@@ -17,27 +18,19 @@ use crate::{
 
 wasm_bindgen_test_configure!(run_in_browser);
 
-fn proto_recv_video_track() -> Track {
-    Track {
-        id: TrackId(1),
-        direction: Direction::Recv {
-            sender: PeerId(234),
-            mid: None,
-        },
-        media_type: MediaType::Video(VideoSettings { is_required: false }),
-    }
-}
-
 #[wasm_bindgen_test]
 async fn on_new_connection_fires() {
     let cons = Connections::default();
 
     let (cb, test_result) = js_callback!(|handle: ConnectionHandle| {
-        cb_assert_eq!(handle.get_remote_id().unwrap(), 234);
+        cb_assert_eq!(
+            handle.get_remote_member_id().unwrap(),
+            "bob".to_string()
+        );
     });
     cons.on_new_connection(cb.into());
 
-    cons.create_connections_from_tracks(PeerId(1), &[proto_recv_video_track()]);
+    cons.create_connection(PeerId(1), &"bob".into());
 
     wait_and_check_test_result(test_result, || {}).await;
 }
@@ -46,9 +39,9 @@ async fn on_new_connection_fires() {
 async fn on_remote_stream_fires() {
     let cons = Connections::default();
 
-    cons.create_connections_from_tracks(PeerId(1), &[proto_recv_video_track()]);
+    cons.create_connection(PeerId(1), &"bob".into());
 
-    let con = cons.get(PeerId(234)).unwrap();
+    let con = cons.get(&"bob".into()).unwrap();
     let con_handle = con.new_handle();
 
     let (cb, test_result) = js_callback!(|stream: RemoteMediaStream| {
@@ -72,9 +65,9 @@ async fn on_remote_stream_fires() {
 async fn tracks_are_added_to_remote_stream() {
     let cons = Connections::default();
 
-    cons.create_connections_from_tracks(PeerId(1), &[proto_recv_video_track()]);
+    cons.create_connection(PeerId(1), &"bob".into());
 
-    let con = cons.get(PeerId(234)).unwrap();
+    let con = cons.get(&"bob".into()).unwrap();
     let con_handle = con.new_handle();
 
     let (tx, rx) = oneshot::channel();
@@ -96,8 +89,8 @@ async fn tracks_are_added_to_remote_stream() {
 #[wasm_bindgen_test]
 async fn on_closed_fires() {
     let cons = Connections::default();
-    cons.create_connections_from_tracks(PeerId(1), &[proto_recv_video_track()]);
-    let con = cons.get(PeerId(234)).unwrap();
+    cons.create_connection(PeerId(1), &"bob".into());
+    let con = cons.get(&"bob".into()).unwrap();
     let con_handle = con.new_handle();
 
     let (on_close, test_result) = js_callback!(|nothing: JsValue| {
@@ -108,4 +101,40 @@ async fn on_closed_fires() {
     cons.close_connection(PeerId(1));
 
     wait_and_check_test_result(test_result, || {}).await;
+}
+
+#[wasm_bindgen_test]
+async fn two_peers_in_one_connection_works() {
+    let cons = Connections::default();
+
+    let (test_tx, mut test_rx) = mpsc::unbounded();
+    let on_new_connection =
+        Closure::wrap(Box::new(move |_: ConnectionHandle| {
+            test_tx.unbounded_send(()).unwrap();
+        }) as Box<dyn Fn(ConnectionHandle)>);
+    cons.on_new_connection(on_new_connection.as_ref().clone().into());
+
+    cons.create_connection(PeerId(1), &"bob".into());
+    test_rx.next().await.unwrap();
+
+    cons.create_connection(PeerId(2), &"bob".into());
+    timeout(300, test_rx.next()).await.unwrap_err();
+}
+
+#[wasm_bindgen_test]
+async fn create_two_connections() {
+    let cons = Connections::default();
+
+    let (test_tx, mut test_rx) = mpsc::unbounded();
+    let on_new_connection =
+        Closure::wrap(Box::new(move |_: ConnectionHandle| {
+            test_tx.unbounded_send(()).unwrap();
+        }) as Box<dyn Fn(ConnectionHandle)>);
+    cons.on_new_connection(on_new_connection.as_ref().clone().into());
+
+    cons.create_connection(PeerId(1), &"bob".into());
+    test_rx.next().await.unwrap();
+
+    cons.create_connection(PeerId(2), &"alice".into());
+    timeout(300, test_rx.next()).await.unwrap();
 }