+++ src/api/client/rpc_connection.rs
@@ -7,9 +7,9 @@ use std::{fmt, time::Duration};
 use actix::Message;
 use derive_more::{From, Into};
 use futures::future::LocalBoxFuture;
-use medea_client_api_proto::{CloseDescription, Command, Event};
+use medea_client_api_proto::{CloseDescription, Command, Event, MemberId};
 
-use crate::{api::control::MemberId, signalling::room::RoomError};
+use crate::signalling::room::RoomError;
 
 /// Newtype for [`Command`] with actix [`Message`] implementation.
 #[derive(Message)]